relocated plotting utility function
it was not being used by anything other than one thing
